+++ src/ProductSaleServices/ProductService.BL/ProductGroupManager.cs
@@ -13,11 +13,13 @@ namespace ProductService.BL
     {
         private readonly IProductRepository _productRepository;
         private readonly IProductGroupRepository _productGroupRepository;
+        private readonly IBlobRepository _blobRepository;
 
-        public ProductGroupManager(IProductRepository productRepository, IProductGroupRepository productGroupRepository)
+        public ProductGroupManager(IProductRepository productRepository, IProductGroupRepository productGroupRepository, IBlobRepository blobRepository)
         {
             _productRepository = productRepository;
             _productGroupRepository = productGroupRepository;
+            _blobRepository = blobRepository;
         }
 
         public async Task<IReadOnlyCollection<ProductGroup>> GetAllProductGroups()
@@ -43,22 +45,31 @@ namespace ProductService.BL
             return await _productGroupRepository.CreateProductGroup(productGroup);
         }
 
-        public async Task<bool> DeleteProductGroup(string id)
+        public async Task<bool> DeleteProductGroup(ProductGroup group)
         {
+            if (group.SampleProduct.IsAvailable == false)
+                return false;
             using (var tran = new TransactionScope(
                 TransactionScopeOption.Required,
                 new TransactionOptions() { IsolationLevel = IsolationLevel.RepeatableRead },
                 TransactionScopeAsyncFlowOption.Enabled))
-            {
-                IReadOnlyCollection<Product> products = await _productRepository.GetProductsByGroupId(id);
+            { 
+                
+                IReadOnlyCollection<Product> products = await _productRepository.GetProductsByGroupId(group.ID);
 
                 foreach (Product product in products)
                 {
+                    for (int i = 0; i < product.PictureLinks.Length; i++)
+                    {
+                        var deleteResult = await _blobRepository.DeleteBlob(product.PictureLinks[i]);
+                        if (deleteResult == false)
+                            return false;
+                    }
                     var res = await _productRepository.DeleteProduct(product.ID);
                     if (res == false)
                         return false;
                 }
-                var result = await _productGroupRepository.DeleteProductGroup(id);
+                var result = await _productGroupRepository.DeleteProductGroup(group.ID);
                 if (result == false)
                     return false;
                 tran.Complete();
@@ -71,6 +82,7 @@ namespace ProductService.BL
             if (oldProductGroup.OwnerID != newProductGroup.OwnerID || oldProductGroup.Quantity != newProductGroup.Quantity)
                 return false;
 
+            // Csak az elérhetőség változhat
             if (oldProductGroup.SampleProduct.ID == newProductGroup.SampleProduct.ID &&
                 oldProductGroup.SampleProduct.GroupID == newProductGroup.SampleProduct.GroupID &&
                 oldProductGroup.SampleProduct.OwnerID == newProductGroup.SampleProduct.OwnerID &&
@@ -88,6 +100,7 @@ namespace ProductService.BL
 
                     foreach (Product product in products)
                     {
+                        // Csak az elérhetőség változhat
                         Product newProduct = new()
                         {
                             ID = product.ID,
@@ -97,6 +110,7 @@ namespace ProductService.BL
                             Description = product.Description,  
                             Name = product.Name,
                             PictureLinks = product.PictureLinks,
+                            EncodedPictures = product.EncodedPictures,
                             IsAvailable = newProductGroup.SampleProduct.IsAvailable,
                         };
                         var res = await _productRepository.UpdateProduct(newProduct.ID, newProduct);
